beatbite_v1/requirements.txt
python-dotenv
newsapi-python
openai
requests
